refactoring astgrep search /replace
